Do not use git to get tests into 1minutetip
When a test is run in 1minutetip, the 1mt machine gets tests
by cloning the db-ci-tests git repository. This is a bit confusing and
impractical (every change in tests has to be pushed to remote git repo
in order to work).

This patch makes run1minutetip.sh copy the db-ci-tests repository
from local machine directly to 1mt machine, without using any remote
repository.